Fix CI and linting errors
This commit addresses several issues identified in the GitHub Actions CI pipeline:

1.  **Backend CI Workflow (`backend-ci.yml`):**
    - I corrected a syntax error by removing an invalid `working-directory` key from the `docker/build-push-action` step.

2.  **Frontend Linting (Unused Variables):**
    - I removed unused imports (`afterEach`, `fireEvent`) in `TagManager.test.jsx`.
    - I removed an unused `textarea` variable assignment in `CommentForm.test.jsx`.
    - I removed unused state variables (`activeTab`, `activities`, `loadingActivities`, `activityError`), the `fetchActivities` function, and related code from `TaskModal.jsx` as the activity display feature was not implemented in this component.
    - I renamed unused error variables to `_e` in `catch` blocks within `TaskCard.jsx`, `CommentItem.jsx`, and `ActivityLogItem.jsx` to satisfy ESLint.

3.  **Frontend Linting (Fast Refresh - `AuthContext.jsx`):**
    - I resolved a Fast Refresh related linting warning by moving the `useAuth` custom hook from `AuthContext.jsx` into its own dedicated file (`frontend/src/contexts/useAuth.jsx`).
    - I updated the import path for `useAuth` in `ProtectedRoute.jsx`.
    - I exported `AuthContext` from `AuthContext.jsx` to allow `useAuth.jsx` to consume it.

These changes should help ensure the CI pipeline runs successfully and improve overall code quality.
