+++ frontend/src/components/ActivityLogItem.jsx
@@ -10,7 +10,7 @@ const ActivityLogItem = ({ activity }) => {
         year: 'numeric', month: 'short', day: 'numeric', 
         hour: '2-digit', minute: '2-digit' 
       });
-    } catch (e) {
+    } catch (_e) {
       return 'Invalid date';
     }
   };
