+++ frontend/src/components/CommentItem.jsx
@@ -10,7 +10,7 @@ const CommentItem = ({ comment }) => {
         year: 'numeric', month: 'short', day: 'numeric', 
         hour: '2-digit', minute: '2-digit' 
       });
-    } catch (e) {
+    } catch (_e) {
       return 'Invalid date';
     }
   };
