+++ frontend/src/components/__tests__/TagManager.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
+import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { server } from '../../mocks/server'; // MSW server
 import { http, HttpResponse } from 'msw'; // MSW for overriding handlers
 import TagManager from '../TagManager';
